+++ Web_api_pizza/Services/AddressService.cs
@@ -1,18 +1,17 @@
 ﻿using System;
 using System.Linq;
 using AutoMapper;
-using CreateDb.Storage;
-using CreateDb.Storage.DTO;
-using CreateDb.Storage.Models;
+using Web_api_pizza.Storage.DTO;
+using Web_api_pizza.Storage.Models;
 
 namespace Web_api_pizza.Services
 { 
     public interface IAddressService
     {
         public AddressDTO GetDeliveryAddress(int id);
-        public void CreateDeliveryAddress(AddressDTO address, int customerId = 0);
-        public void EditDeliveryAddress(AddressDTO address);
-        public void RemoveDeliveryAddress(int id);
+        public string CreateDeliveryAddress(AddressDTO address, int customerId = 0);
+        public string EditDeliveryAddress(AddressDTO address);
+        public string RemoveDeliveryAddress(int id);
     }
     public class AddressService : IAddressService
     {
@@ -26,25 +25,28 @@ namespace Web_api_pizza.Services
 
         public AddressDTO GetDeliveryAddress(int id)
         {
-            //try
-            //{
                 var addressEntity = _context.Addresses.FirstOrDefault(a => a.Id == id);
                 var addressDTO = _mapper.Map<AddressEntity, AddressDTO>(addressEntity);
                 return addressDTO;
-            //}
-            //catch
-            //{
-            //    return "Неверные данные пользователя";
-            //}
         }
-        public void CreateDeliveryAddress(AddressDTO address, int customerId = 0)
+        public string CreateDeliveryAddress(AddressDTO address, int customerId = 0)
         {
+            address.Id = 0;
+            string message;
             var checkAddress = FindAddress(address);
             if (checkAddress == null)
             {
                 var addressEntity = _mapper.Map<AddressDTO, AddressEntity>(address);
                 _context.Addresses.Add(addressEntity);
                 _context.SaveChanges();
+                message = "Адрес добавлен";
+            }
+            //ненужное сообщение для клиента
+            //только для проверки
+            //удалить после тестирования
+            else
+            { 
+                message = "Адрес уже есть в базе";
             }
             if(customerId != 0)
             {
@@ -57,8 +59,20 @@ namespace Web_api_pizza.Services
                     var customerAddress = new CustomerAddressEntity { AddressEntityId = findAddress.Id, CustomerEntityId = customerId };
                     _context.CustomerAddressEntities.Add(customerAddress);
                     _context.SaveChanges();
+                    //ненужное сообщение для клиента
+                    //только для проверки
+                    //удалить после тестирования
+                    message += "\nСвязь между пользователем и адресом создана";
+                }
+                //ненужное сообщение для клиента
+                //только для проверки
+                //удалить после тестирования
+                else
+                {
+                    message += "\nСвязь между пользователем и адресом уже существует";
                 }
             }
+            return message;
         }
         private AddressEntity FindAddress(AddressDTO address)
         {
@@ -71,10 +85,24 @@ namespace Web_api_pizza.Services
                 .FirstOrDefault();
             return findAddress;
         }
-        public void EditDeliveryAddress(AddressDTO address)
+        public string EditDeliveryAddress(AddressDTO address)
         {
+            string message;
             var editableAddress = _context.Addresses.FirstOrDefault(a => a.Id == address.Id);
-
+            if(editableAddress == null)
+            {
+                message = null;
+                return message;
+            }
+            if(editableAddress.City == address.City &&
+            editableAddress.Street == address.Street &&
+            editableAddress.NumberOfBuild == address.NumberOfBuild &&
+            editableAddress.NumberOfEntrance == address.NumberOfEntrance &&
+            editableAddress.Apartment == address.Apartment)
+            {
+                message = "что изменять то?";
+                return message;
+            }
             editableAddress.City = address.City;
             editableAddress.Street = address.Street;
             editableAddress.NumberOfBuild = address.NumberOfBuild;
@@ -82,12 +110,22 @@ namespace Web_api_pizza.Services
             editableAddress.Apartment = address.Apartment;
 
             _context.SaveChanges();
+            message = "Адрес изменен";
+            return message;
         }
-        public void RemoveDeliveryAddress(int id)
+        public string RemoveDeliveryAddress(int id)
         {
             var removableAddress = _context.Addresses.FirstOrDefault(a => a.Id == id);
+            string message;
+            if(removableAddress == null)
+            {
+                message = null;
+                return message;
+            }
             _context.Addresses.Remove(removableAddress);
             _context.SaveChanges();
+            message = "Адрес удален";
+            return message;
         }
     }
 }
